# docs/swagger.yaml
basePath: /api/v1
definitions:
  ErrorResponse:
    properties:
      error:
        example: invalid_request
        type: string
      message:
        example: The request was invalid
        type: string
      status:
        example: 400
        type: integer
    type: object
...
  SuccessResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      status:
        example: 200
        type: integer
    type: object
...
  github_com_the-monkeys_monkeys-identity_internal_models.Organization:
    properties:
      billing_tier:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: string
      max_resources:
        type: integer
      max_users:
        type: integer
      metadata:
        description: JSONB as string
        type: string
      name:
        type: string
      parent_id:
        type: string
      settings:
        description: JSONB as string
        type: string
      slug:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_handlers.updateSettingsRequest:
    properties:
      settings:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@themonkeys.com
    name: The Monkeys Team
    url: https://github.com/the-monkeys
  description: A comprehensive IAM system providing authentication, authorization,
    and access control for multi-tenant organizations with policy-based security,
    session management, and comprehensive audit trails.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://themonkeys.com/terms
  title: Monkeys Identity & Access Management API
  version: "1.0"
paths:
...
  /admin/settings:
    get:
      description: Retrieve system-wide global settings
      produces:
      - application/json
      responses:
        "200":
          description: Global settings retrieved
          schema:
            $ref: '#/definitions/SuccessResponse'
      security:
      - BearerAuth: []
      summary: Get global settings
      tags:
      - System Administration
    put:
      consumes:
      - application/json
      description: Update system-wide global settings
      produces:
      - application/json
      responses:
        "200":
          description: Global settings updated
          schema:
            $ref: '#/definitions/SuccessResponse'
      security:
      - BearerAuth: []
      summary: Update global settings
      tags:
      - System Administration
...
  /organizations:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of organizations
      parameters:
      - description: Items per page (1-1000)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Organizations retrieved
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid pagination parameters
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: List organizations
      tags:
      - Organization Management
    post:
      consumes:
      - application/json
      description: Create a new organization (tenant)
      parameters:
      - description: Organization details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_the-monkeys_monkeys-identity_internal_models.Organization'
      produces:
      - application/json
      responses:
        "201":
          description: Organization created
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Organization already exists
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create organization
      tags:
      - Organization Management
  /organizations/{id}:
    delete:
      description: Soft delete an organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Organization deleted
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid organization ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete organization
      tags:
      - Organization Management
    get:
      description: Retrieve an organization by ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Organization retrieved
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid organization ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get organization
      tags:
      - Organization Management
    put:
      consumes:
      - application/json
      description: Update organization attributes
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated organization
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_the-monkeys_monkeys-identity_internal_models.Organization'
      produces:
      - application/json
      responses:
        "200":
          description: Organization updated
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update organization
      tags:
      - Organization Management
  /organizations/{id}/groups:
    get:
      description: List all groups belonging to an organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Groups retrieved
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid organization ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: List organization groups
      tags:
      - Organization Management
  /organizations/{id}/policies:
    get:
      description: List all policies within an organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Policies retrieved
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid organization ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: List organization policies
      tags:
      - Organization Management
  /organizations/{id}/resources:
    get:
      description: List all resources owned by an organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Resources retrieved
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid organization ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: List organization resources
      tags:
      - Organization Management
  /organizations/{id}/roles:
    get:
      description: List all roles defined in an organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Roles retrieved
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid organization ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: List organization roles
      tags:
      - Organization Management
  /organizations/{id}/settings:
    get:
      description: Retrieve settings JSON for an organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Settings retrieved
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid organization ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get organization settings
      tags:
      - Organization Management
    put:
      consumes:
      - application/json
      description: Update settings JSON for an organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated settings payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.updateSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Settings updated
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Organization not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update organization settings
      tags:
      - Organization Management
  /organizations/{id}/users:
    get:
      description: List all users belonging to an organization
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Invalid organization ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - BearerAuth: []
      summary: List organization users
      tags:
      - Organization Management
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"